feat | update task controller and routes

+++ api/app/Http/Controllers/TaskController.php
@@ -14,29 +14,26 @@ class TaskController extends Controller
 		return $user->tasks;
 	}
 
-	// public function show(Task $task){
-	// 	return $task;
-	// }
-
 	public function create(Request $request){
 		$user = auth()->guard('api')->user();
 
 		$this->validate($request, [
-			'title' => 'required|max:191',
-			'checked' => 'required'
+			'title' => 'required|max:191'
 		]);
-
-		$task = $user->tasks()->create($request->all());
+		$task = $user->tasks()->create([
+			'title' => $request->title,
+			'checked' => 0
+		]);
 		return response()->json($task, 201);
 	}
 
 	public function update(Request $request, Task $task){
-		$task = Task::findOrFail($task->id);
+		// $task = Task::findOrFail($task->id);
 		$user = auth()->guard('api')->user();
 		if($user->id !== $task->user_id){
 			throw new AuthenticationException();
 		}
-		
+
 		$this->validate($request, [
 			'title' => 'required|max:191',
 			'checked' => 'required'
@@ -48,6 +45,12 @@ class TaskController extends Controller
 	}
 
 	public function delete(Task $task){
+		// $task = Task::findOrFail($taskId);
+		$user = auth()->guard('api')->user();
+		if($user->id !== $task->user_id){
+			throw new AuthenticationException();
+		}
+
 		$task->delete();
 
 		return response()->json(null, 204);
